agents/VH_Resource_Agent/requirements.txt
uagents
python-dotenv
requests
openai
qdrant-client
pydantic
